a little update.
this update, I've implemented some math to determine which points of the
object should be in the calculation later.
and some toggles to see the levels of the program.
also implemented some guiding lines to help visualize the position of
the shadow, where it should be.

SPACE - turn on/off flashlight.
L - turn on/off guide lines.
M - turn on/off background image.
C - turn on/off cursor.